2006-08-22  Sebastien Pouliot  <sebastien@ximian.com>
	* gdipFunctions.cs: Fix GdipGetImageWidth to use an uint for it's
	width out parameter.
	* Image.cs: Cast the uint into an int for the Width property.


svn path=/trunk/mcs/; revision=64211

+++ mcs/class/System.Drawing/System.Drawing/Image.cs
@@ -730,11 +730,11 @@ public abstract class Image : MarshalByRefObject, IDisposable , ICloneable, ISer
 	[DesignerSerializationVisibility (DesignerSerializationVisibility.Hidden)]
 	public int Width {
 		get {
-			int width;			
+			uint width;			
 			Status status = GDIPlus.GdipGetImageWidth (nativeObject, out width);		
 			GDIPlus.CheckStatus (status);			
 			
-			return width;
+			return (int)width;
 		}
 	}
